backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "team13"}
]

requires-python = ">=3.12"
# --- 主依赖：只保留运行应用所必需的包 ---
dependencies = [
    "sqlmodel (>=0.0.24,<0.0.25)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "cryptography (>=42.0.2,<43.0.0)",
    "twisted (>=24.7.0rc1,<25.0.0)",
    "urllib3 (>=2.5.0,<3.0.0)",
    "pyjwt (>=2.4.0,<3.0.0)",
]

# --- 可选依赖/开发依赖：用于测试的包放在这里 ---
[project.optional-dependencies]
test = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)", # 使用较新的版本，并删除了冲突的旧版本
    "httpx (>=0.28.1,<0.29.0)",      # 使用较新的版本
]

[tool.poetry]
packages = [
    { include = "app" },
]

[tool.poetry.group.dev.dependencies]
httpx = "^0.28.1"
pytest-asyncio = "^1.2.0"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"